feat: Export fetchPolicies directly
This way we can import only fetchPolicies if needed
