+++ packages/cozy-client/src/CozyClient.js
@@ -24,6 +24,7 @@ import {
   getCollectionFromState,
   getDocumentFromState
 } from './store'
+import fetchPolicies from './policies'
 import Schema from './Schema'
 import { chain } from './CozyLink'
 import ObservableQuery from './ObservableQuery'
@@ -1108,37 +1109,7 @@ class CozyClient {
   }
 }
 
-/**
- * Use those fetch policies with `<Query />` to limit the number of re-fetch.
- *
- * @example
- * ```
- * const olderThan30s = CozyClient.fetchPolicies.olderThan(30 * 1000)
- * <Query fetchPolicy={olderThan30s} />
- * ```
- */
-CozyClient.fetchPolicies = {
-  /**
-   * Returns a fetchPolicy that will only re-fetch queries that are older
-   * than `<delay>` ms.
-   *
-   * @param  {number} delay - Milliseconds since the query has been fetched
-   * @returns {Function} Fetch policy to be used with `<Query />`
-   */
-  olderThan: delay => queryState => {
-    if (!queryState || !queryState.lastUpdate) {
-      return true
-    } else {
-      const elapsed = Date.now() - queryState.lastUpdate
-      return elapsed > delay
-    }
-  },
-
-  /**
-   * Fetch policy that deactivates any fetching.
-   */
-  noFetch: () => false
-}
+CozyClient.fetchPolicies = fetchPolicies
 
 MicroEE.mixin(CozyClient)
